added automatic silent auth on session restore (opt-in)
