+++ addon/authenticators/auth0-base.js
@@ -1,6 +1,10 @@
 import RSVP from 'rsvp';
+import { get } from '@ember/object';
+import { isEmpty } from '@ember/utils';
 import { inject as service } from '@ember/service';
 import BaseAuthenticator from 'ember-simple-auth/authenticators/base';
+import { Auth0Error } from '../utils/errors'
+import createSessionDataObject from '../utils/create-session-data-object';
 import getSessionExpiration from '../utils/get-session-expiration';
 import now from '../utils/now';
 
@@ -10,8 +14,40 @@ export default BaseAuthenticator.extend({
     const expiresAt = getSessionExpiration(data || {});
     if(expiresAt > now()) {
       return RSVP.resolve(data);
+    } else if(get(this, 'auth0.silentAuthOnSessionRestore')) {
+      const options = get(this, 'auth0.silentAuthOptions');
+      return this._performSilentAuth(options)
     } else {
       return RSVP.reject();
     }
   },
+
+  // resolve the data returned by a parseHash/silentAuth result.
+  _resolveAuthResult(authResult, resolve, reject) {
+    if (isEmpty(authResult)) {
+      reject();
+    }
+    const auth0 = get(this, 'auth0').getAuth0Instance();
+    const getUserInfo = auth0.client.userInfo.bind(auth0.client);
+
+    getUserInfo(authResult.accessToken, (err, profile) => {
+      if (err) {
+        return reject(new Auth0Error(err));
+      }
+
+      resolve(createSessionDataObject(profile, authResult));
+    });
+  },
+
+  // performs silent authentication & handles the result in a promise.
+  _performSilentAuth(options) {
+    return new RSVP.Promise((resolve, reject) => {
+      // perform silent auth via auth0's checkSession function (called in the service);
+      // if successful, use the same logic as the url-hash authenticator since the
+      // result of checkSession is the same as parseHash.
+      get(this, 'auth0').silentAuth(options).then(authenticatedData => {
+        this._resolveAuthResult(authenticatedData, resolve, reject);
+      }, reject);
+    });
+  }
 });
